preserve focusedIndex on consecutive selections when using lazy loading
When using a data provider (lazy loading) we need to store the value of the `_focusedIndex`
in order to retain the overlay scroll position after the value is reset
(reseting the value sets the  to -1).
This ensures that on consecutive value selections, the overlay is opened
at the correct position in the list of items.