fix focus ring bug when preserving focused index
When using a data provider we need to manually scroll into view
and reset the value of the  to remove the focus
ring from the last selected item.
This ensures that on the first consecutive opening, the
overlay is opened at the correct position in the list of items
without the focus ring on the last selected item.

for more info see: https://github.com/gatanaso/multiselect-combo-box-flow/issues/26