allow inputing of custom values
- Introduced the `allows-custom-values` attribute, which enables the
multiselect-combo-box component to accept custom input values.
If enabled, when the user inputs a value that is not present in the existing
list of items, a 'custom-values-set' event is fired with the value.
It is up to the end user to decide what they will do with this value as it
is not automatically added to the list of items or values